refactor(handler): merge eval_batch.go into eval.go
Single eval surface: handlers, EvaluateBatch, GET decode/validate, engine.
